write result of createJob mutation to cache
Normally, createJob makes 2 graphql requests, 1 for creatJob mutation and the other to query for the newly created job. Since createJob already returns the newly created job, the second graphql request is not needed. To avoid the second unnecessary graphql request, write the result of the createJob mutation to cache, passing the job query (not createJob) as the query arg.